Created external file for form model and rearranged project tree
