+++ config.py
@@ -1,0 +1,4 @@
+# /config.py
+
+class Config:
+    SECRET_KEY = '11122233344'
